Update calendar-generator.js
The "daysInMonth"  is being set to the number of days from the previous month causing calendar.html to display the days in each month incorrectly.

Creating a new date with the day parameter as 0 causes the date is set to the maximum day of the pervious month. This is fixed by incrementing month by 1.